use Ducks pattern
