--- src/store/actionTypes.js
@@ -1,2 +1,0 @@
-export const taskUpdated = 'task/updated';
-export const taskDelete = 'task/delete';
