--- src/store/actions.js
@@ -1,22 +1,0 @@
-import * as actionTypes from './actionTypes';
-
-export function taskCompleted(id) {
-  return {
-    type: actionTypes.taskUpdated,
-    payload: { id, completed: true },
-  };
-}
-
-export function titleChanged(id) {
-  return {
-    type: actionTypes.taskUpdated,
-    payload: { id, title: `New title for ${id}` },
-  };
-}
-
-export function taskDeleted(id) {
-  return {
-    type: actionTypes.taskDelete,
-    payload: { id: id },
-  };
-}
